Diversify LinearGradient() IIFE for other sections
The LinearGradient() IIFE was only being used for generating a linear
gradient for the daily forecast section.
However, it will now also be used for generating a linear gradient for
the uv-index section.
Therefore, the IIFE has been diversified and the functions to generate
linear gradient of daily forecast and uv-index sections have been
separated.