Remove avg temp & add marker to daily forecast LG
The average temperature will no longer be displayed in the daily
forecast linear gradient (LG).
Further, a marker has been added for the current day's linear gradient
to show where the current temperature lies on the linear gradient scale
between min and max temperatures.
The marker will only be displayed on the scale of the current day as it
serves as the current temperature.