Create an object to store both of fetched results
Since the result we fetch should be of two types—with imperial units and
with metric units—we need to store the data somewhere so that we can
instantly switch the data when the user desires to view it in the other
unit system.
Storing this data would save us the time of fetching it for the user
repeatedly and we would be able to seamlessly transition between the two
types of data on display.